refactor: Move Authentication config from product module to authentication module
