+++ liberty-authentication-library/src/main/java/com/liberty52/authentication/web/configurer/AuthenticationWebConfiguration.java
@@ -1,0 +1,31 @@
+package com.liberty52.authentication.web.configurer;
+
+import com.liberty52.authentication.web.filter.AuthenticationFilter;
+import com.liberty52.authentication.web.resolver.UserContextArgumentResolver;
+import jakarta.servlet.Filter;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+public class AuthenticationWebConfiguration implements WebMvcConfigurer {
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new UserContextArgumentResolver());
+    }
+
+    @Bean
+    @SuppressWarnings("rawtypes")
+    public FilterRegistrationBean addAuthenticationFilter() {
+        FilterRegistrationBean<Filter> filterRegistrationBean = new FilterRegistrationBean<>();
+        filterRegistrationBean.setFilter(new AuthenticationFilter());
+        filterRegistrationBean.setOrder(1);
+        filterRegistrationBean.addUrlPatterns("/*");
+        return filterRegistrationBean;
+    }
+}
