+++ src/main/java/com/liberty52/product/global/config/WebConfig.java
@@ -1,18 +1,10 @@
 package com.liberty52.product.global.config;
 
-import com.liberty52.authentication.web.filter.AuthenticationFilter;
-import com.liberty52.authentication.web.resolver.UserContextArgumentResolver;
 import com.liberty52.product.global.constants.TranslationConstants;
-import jakarta.servlet.Filter;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
@@ -22,25 +14,9 @@ public class WebConfig implements WebMvcConfigurer {
         registry.addConverter(new TranslationConstants.LangCodeConverter());
     }
 
-
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(new RequestLoggingInterceptor());
-    }
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new UserContextArgumentResolver());
-    }
-
-    @Bean
-    @SuppressWarnings("rawtypes")
-    public FilterRegistrationBean addAuthenticationFilter() {
-        FilterRegistrationBean<Filter> filterRegistrationBean = new FilterRegistrationBean<>();
-        filterRegistrationBean.setFilter(new AuthenticationFilter());
-        filterRegistrationBean.setOrder(1);
-        filterRegistrationBean.addUrlPatterns("/*");
-        return filterRegistrationBean;
     }
 }
